fix janna bug
For some reason janna's json response changed range's value to a long integer. Since it was being mapped to a List<Int>, a NumberFormatException was occurred.

+++ app/src/main/java/com/machado001/lilol/rotation/model/dto/SpellDto.kt
@@ -17,7 +17,7 @@ data class SpellDto(
     val vars: List<Any>,
     val costType: String,
     val maxammo: String,
-    val range: List<Int>,
+    val range: List<Long>,
     val rangeBurn: String,
     val image: ImageDto,
     val resource: String
